r281/sofa-dev : Modify Intersection class to be a base class of real intersection methods. Add 3 (partial) implementations: DiscreteIntersection, ContinuousIntersection, ProximityIntersection VoxelGrid now correctly compute BBoxes depending on the method used. Modify demo6Triangle to use Proximity.
 //partial sync (23/30 changes)//

+++ src/Sofa/Components/GL/OglModel.cpp
@@ -291,9 +291,9 @@ void OglModel::applyScale(double scale)
 {
     for (int i = 0; i < nbVertices; i++)
     {
-        vertices[i * 3] *= scale;
-        vertices[i * 3 + 1] *= scale;
-        vertices[i * 3 + 2] *= scale;
+        vertices[i * 3] *= (float)scale;
+        vertices[i * 3 + 1] *= (float)scale;
+        vertices[i * 3 + 2] *= (float)scale;
     }
 }
